+++ be/src/common/config.h
@@ -829,6 +829,8 @@ CONF_Int32(doris_remote_scanner_thread_pool_thread_num, "16");
 // number of s3 scanner thread pool queue size
 CONF_Int32(doris_remote_scanner_thread_pool_queue_size, "10240");
 
+CONF_Double(s2geo_eps, "0.000000001");
+
 #ifdef BE_TEST
 // test s3
 CONF_String(test_s3_resource, "resource");