+++ be/src/util/brpc_client_cache.h
@@ -65,7 +65,7 @@ public:
 
     std::shared_ptr<T> get_client(const std::string& host_port) {
         std::shared_ptr<T> stub_ptr;
-        auto get_value = [&stub_ptr](typename StubMap<T>::mapped_type& v) { stub_ptr = v; };
+        auto get_value = [&stub_ptr](const auto& v) { stub_ptr = v.second; };
         if (LIKELY(_stub_map.if_contains(host_port, get_value))) {
             return stub_ptr;
         }
@@ -73,7 +73,7 @@ public:
         // new one stub and insert into map
         auto stub = get_new_client_no_cache(host_port);
         _stub_map.try_emplace_l(
-                host_port, [&stub](typename StubMap<T>::mapped_type& v) { stub = v; }, stub);
+                host_port, [&stub](const auto& v) { stub = v.second; }, stub);
         return stub;
     }